third_party/blink/web_tests/http/tests/security/only-top-level-navigation-hsts-upgrade-expected.txt

CONSOLE MESSAGE: iframe successfully loaded via http
CONSOLE MESSAGE: HSTS iframe successfully didn't load
CONSOLE MESSAGE: Window's origin is: https://hsts-example.test:8443
